requirements.txt
streamlit
mysql-connector-python
pandas
plotly
openpyxl
mysql-connector-python pooling
streamlit-aggrid
